完成 4.序列化器与SPI机制 的基本设计

+++ example-common/src/main/java/com/wb/example/common/service/UserService.java
@@ -15,7 +15,7 @@ public interface UserService {
      */
     User getUser(User user);
 
-    default short getNumber(){
+    default int getNumber(){
         return 1;
     }
 }
