+++ example-provider/src/main/java/com/wb/example/provider/service/impl/UserServiceImpl.java
@@ -13,7 +13,7 @@ public class UserServiceImpl implements UserService {
         return user;
     }
 
-    public short getNumber(){
+    public int getNumber(){
         return 55;
     }
 }
